Making performance improvements to SSH connections
* Removing unneeded SSH _connect() call

By removing this SSH _connect() call performance is improved.
_create_ssh() only needs to setup the SSH object.  The open()
function handles calling _connect() already so this one was
not necessary.

* Added logic to check if a SSH transport is active

Previously it was only checking if the transport layer was
missing but there is a possibility the transport layer could
exist and not be active.  This change will add checks for this.

Change-Id: Ie5d21ed28744d9a0683fbac4e55c859917e9416d