Adding more descriptive debug messages to client __init__
Errors occuring during the client __init__ call were being
described very generically. This patch adds more details so that
it is easier to debug an initialization issue.

Also, added the ability to toggle debug on before the client is
initialized.  This will allow the HTTP requests that happen
during __init__ to be recorded.

bug reference:
https://github.com/hp-storage/python-3parclient/issues/9

Change-Id: I896fa9ebe422d9520980370aab2ae37a70322eca