Removing 'active' VLUN check from getVLUN call
There was a check for 'active' VLUNs during the getVLUN call when
using newer firmware on a backend. The check for 'active' is not
needed and was causing a unit test failure when run on backends
with newer firmware. This also keeps the API consistent between
newer and older firmware versions.


Change-Id: Ifd3223d92930c76645762f9fc956bd28df88d5fe