Less nargs/more shlex to fix fake SSH usage error
The paramiko fake SSH server was not properly splitting
args that were quoted and contained spaces (comments).
It was working due to luck of order and also in one case
handling the split comment like repeated args.

The error was inconsistent, but it seems to be failing more
now that we run with PY34.

shlex.split() knows how to do posix-like splitting of the
command and the nargs option was removed.

Change-Id: I5d2138606e3e6905166296a307d4af9ca89c9c7e