Add py36 and py37 support in tox
This patch adds the python 3.6 and python 3.7 designations in tox tests.

I ran python 37 tests here:
http://paste.openstack.org/show/5TPPZ7E0yIj1K4iyLCEZ/

We need this patch to merge to get openstack upstream global requirments
patch to land so we can add openstack cinder to allow optionally
automatically installing this client lib for users that want a container
with 3par driver support built.