Fix debug logging
We currently have a couple of issues with the enabling and disabling of
debug logs:

- Passing "debug=True" on client creation will not output ssh logs.

- Once enabled, SSH and HTTP debug logs cannot be disabled.

This patch fixes the first issue by making "set_debug_flag" as class
methods and always calling ssh's "set_debug_flag", unlike now that we
only call it when the ssh instance has already been created.

For the second issue we add the missing disabling code in both
"set_debug_flag" methods.

Closes #76